Various fixes for amd64 compilation errors under gcc.  Also patch __init__.py to try setting dlopenflags even if the dl module isn't found.  The latter probably won't help but shouldn't hurt.
git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1949 7152ba15-5bd8-0310-b828-e896570a432b

+++ csp/cspsim/f16/MultiFunctionDisplay.cpp
@@ -223,7 +223,7 @@ void DisplayText::setInvert(bool invert) {
 void DisplayText::setText(std::string const &text, bool invert) {
 	TokenQueue tokens(text, "\n");
 	m_Lines.reserve(tokens.size());
-	unsigned widest = 0;
+	size_t widest = 0;
 	for (unsigned i = 0; i < tokens.size(); ++i) {
 		widest = std::max(tokens[i].size(), widest);
 		display::ElementText *text = (i < m_Lines.size()) ? m_Lines[i].get() : 0;
